backend/requirements.txt
numpy==1.26.4
pandas
cython
bcrypt==4.0.1
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pydantic
scikit-learn
scikit-surprise
python-dotenv
httpx
boto3
apscheduler
